Split interfaces into separate files
